Updated Stat Cards
Now with half the memory usage!

+++ game/systems/construction.js
@@ -1,5 +1,4 @@
 import { ConstructionAction } from "../actions/constructionAction.js";
-import { DefaultTypes } from "../defaultTypes.js";
 import { ArmyEntity } from "../init/armyEntity.js";
 import { SpawnSystem } from "./spawn.js";
 
@@ -7,25 +6,6 @@ import { SpawnSystem } from "./spawn.js";
  * Collection of functions revolving around the construction of buildings.
  */
 export const ConstructionSystem = function() {}
-
-/**
- * Returns a spawnable entity config.
- * 
- * @param {*} gameContext 
- * @param {*} entity 
- * @returns {SpawnConfigType}
- */
-const getResult = function(gameContext, entity) {
-    const { world } = gameContext;
-    const { turnManager } = world;
-    const { tileX, tileY } = entity.getComponent(ArmyEntity.COMPONENT.POSITION);
-    const { teamID } = entity.getComponent(ArmyEntity.COMPONENT.TEAM);
-    const entityID = entity.getID();
-    const owners = turnManager.getOwnersOf(entityID).map(actor => actor.getID());
-    const type = entity.config.constructionResult;
-    
-    return DefaultTypes.createSpawnConfig(type, teamID, owners, tileX, tileY);
-}
 
 /**
  * Creates a build request for the entity or finishes up construction.
@@ -48,10 +28,14 @@ ConstructionSystem.onInteract = function(gameContext, entity, actorID) {
     
     if(entity.isConstructionComplete()) {
         if(!actionQueue.isRunning()) {
-            const result = getResult(gameContext, entity);
+            const willBuild = confirm(`Finish construction?`);
 
-            SpawnSystem.destroyEntity(gameContext, entity);
-            SpawnSystem.createEntity(gameContext, result);
+            if(willBuild) {
+                const result = entity.getConstructionResult(gameContext);
+
+                SpawnSystem.destroyEntity(gameContext, entity);
+                SpawnSystem.createEntity(gameContext, result);
+            }
         }
     } else {
         const entityID = entity.getID();
